scripts/requirements.txt
mysql-connector-python>=8.0
psycopg2-binary>=2.9
python-dotenv>=1.0
